Revert "test: enable URL scheme validation via MockUIRule"
This reverts commit 8787155841efd36b358b0096e47cb3b36f7473b8.
