+++ vaadin-charts-flow-parent/vaadin-charts-flow/src/test/java/com/vaadin/flow/component/charts/model/CreditsTest.java
@@ -8,38 +8,32 @@
  */
 package com.vaadin.flow.component.charts.model;
 
-import org.junit.Assert;
-import org.junit.Rule;
-import org.junit.Test;
-
-import com.vaadin.tests.MockUIRule;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 public class CreditsTest {
 
-    @Rule
-    public MockUIRule mockUIRule = new MockUIRule();
-
     @Test
-    public void setHref_safeScheme_hrefSet() {
+    void setHref_safeScheme_hrefSet() {
         Credits credits = new Credits();
         credits.setHref("https://vaadin.com");
 
-        Assert.assertEquals("https://vaadin.com", credits.getHref());
+        Assertions.assertEquals("https://vaadin.com", credits.getHref());
     }
 
     @Test
-    public void setHref_unsafeScheme_throws() {
+    void setHref_unsafeScheme_throws() {
         Credits credits = new Credits();
 
-        Assert.assertThrows(IllegalArgumentException.class,
+        Assertions.assertThrows(IllegalArgumentException.class,
                 () -> credits.setHref("javascript:alert(1)"));
     }
 
     @Test
-    public void setUnsafeHref_unsafeScheme_hrefSet() {
+    void setUnsafeHref_unsafeScheme_hrefSet() {
         Credits credits = new Credits();
         credits.setUnsafeHref("javascript:alert(1)");
 
-        Assert.assertEquals("javascript:alert(1)", credits.getHref());
+        Assertions.assertEquals("javascript:alert(1)", credits.getHref());
     }
 }
