+++ vaadin-charts-flow-parent/vaadin-charts-flow/src/test/java/com/vaadin/flow/component/charts/model/ExportingTest.java
@@ -8,64 +8,58 @@
  */
 package com.vaadin.flow.component.charts.model;
 
-import org.junit.Assert;
-import org.junit.Rule;
-import org.junit.Test;
-
-import com.vaadin.tests.MockUIRule;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 public class ExportingTest {
 
-    @Rule
-    public MockUIRule mockUIRule = new MockUIRule();
-
     @Test
-    public void setUrl_safeScheme_urlSet() {
+    void setUrl_safeScheme_urlSet() {
         Exporting exporting = new Exporting();
         exporting.setUrl("https://export.highcharts.com");
 
-        Assert.assertEquals("https://export.highcharts.com",
+        Assertions.assertEquals("https://export.highcharts.com",
                 exporting.getUrl());
     }
 
     @Test
-    public void setUrl_unsafeScheme_throws() {
+    void setUrl_unsafeScheme_throws() {
         Exporting exporting = new Exporting();
 
-        Assert.assertThrows(IllegalArgumentException.class,
+        Assertions.assertThrows(IllegalArgumentException.class,
                 () -> exporting.setUrl("javascript:alert(1)"));
     }
 
     @Test
-    public void setUnsafeUrl_unsafeScheme_urlSet() {
+    void setUnsafeUrl_unsafeScheme_urlSet() {
         Exporting exporting = new Exporting();
         exporting.setUnsafeUrl("javascript:alert(1)");
 
-        Assert.assertEquals("javascript:alert(1)", exporting.getUrl());
+        Assertions.assertEquals("javascript:alert(1)", exporting.getUrl());
     }
 
     @Test
-    public void setLibURL_safeScheme_libURLSet() {
+    void setLibURL_safeScheme_libURLSet() {
         Exporting exporting = new Exporting();
         exporting.setLibURL("https://code.highcharts.com/lib");
 
-        Assert.assertEquals("https://code.highcharts.com/lib",
+        Assertions.assertEquals("https://code.highcharts.com/lib",
                 exporting.getLibURL());
     }
 
     @Test
-    public void setLibURL_unsafeScheme_throws() {
+    void setLibURL_unsafeScheme_throws() {
         Exporting exporting = new Exporting();
 
-        Assert.assertThrows(IllegalArgumentException.class,
+        Assertions.assertThrows(IllegalArgumentException.class,
                 () -> exporting.setLibURL("javascript:alert(1)"));
     }
 
     @Test
-    public void setUnsafeLibURL_unsafeScheme_libURLSet() {
+    void setUnsafeLibURL_unsafeScheme_libURLSet() {
         Exporting exporting = new Exporting();
         exporting.setUnsafeLibURL("javascript:alert(1)");
 
-        Assert.assertEquals("javascript:alert(1)", exporting.getLibURL());
+        Assertions.assertEquals("javascript:alert(1)", exporting.getLibURL());
     }
 }
