+++ vaadin-flow-components-shared-parent/vaadin-flow-components-test-util/src/main/java/com/vaadin/tests/MockUIRule.java
@@ -17,7 +17,6 @@ package com.vaadin.tests;
 
 import java.util.List;
 import java.util.Locale;
-import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Future;
 import java.util.concurrent.locks.Lock;
@@ -52,12 +51,6 @@ public class MockUIRule extends ExternalResource {
         DeploymentConfiguration deploymentConfig = Mockito
                 .mock(DeploymentConfiguration.class);
         Mockito.when(deploymentConfig.isProductionMode()).thenReturn(false);
-        // Enable URL-scheme validation (e.g. UrlUtil.isSafeUrl) during tests by
-        // exposing the default set of safe schemes through the deployment
-        // configuration. Without this stub the mock would return null and any
-        // scheme validation would fail.
-        Mockito.when(deploymentConfig.getUrlSafeSchemes())
-                .thenReturn(Set.of("http", "https", "mailto", "tel", "ftp"));
         Mockito.when(service.getDeploymentConfiguration())
                 .thenReturn(deploymentConfig);
 
@@ -68,7 +61,6 @@ public class MockUIRule extends ExternalResource {
 
         UI.setCurrent(ui);
         VaadinSession.setCurrent(session);
-        VaadinService.setCurrent(service);
     }
 
     @Override
@@ -77,7 +69,6 @@ public class MockUIRule extends ExternalResource {
         removeAll();
         UI.setCurrent(null);
         VaadinSession.setCurrent(null);
-        VaadinService.setCurrent(null);
     }
 
     /**
