+++ vaadin-side-nav-flow-parent/vaadin-side-nav-flow/src/test/java/com/vaadin/flow/component/sidenav/tests/SideNavItemTest.java
@@ -139,10 +139,29 @@ public class SideNavItemTest {
     }
 
     @Test
+    public void setPathWithUnsafeScheme_throws() {
+        Assert.assertThrows(IllegalArgumentException.class,
+                () -> sideNavItem.setPath("javascript:alert(1)"));
+    }
+
+    @Test
     public void setUnsafePathWithUnsafeScheme_pathSet() {
         sideNavItem.setUnsafePath("javascript:alert(1)");
 
         assertPath("javascript:alert(1)");
+    }
+
+    @Test
+    public void constructor_labelPath_unsafeScheme_throws() {
+        Assert.assertThrows(IllegalArgumentException.class,
+                () -> new SideNavItem("Docs", "javascript:alert(1)"));
+    }
+
+    @Test
+    public void constructor_labelPathPrefixComponent_unsafeScheme_throws() {
+        Assert.assertThrows(IllegalArgumentException.class,
+                () -> new SideNavItem("Docs", "javascript:alert(1)",
+                        new Div()));
     }
 
     @Test
